+++ apps/extension/package.json
@@ -23,7 +23,6 @@
     "@penumbra-labs/registry": "^7.0.0",
     "@penumbra-zone/bech32m": "workspace:*",
     "@penumbra-zone/client": "workspace:*",
-    "@penumbra-zone/constants": "workspace:*",
     "@penumbra-zone/crypto-web": "workspace:*",
     "@penumbra-zone/perspective": "workspace:*",
     "@penumbra-zone/query": "workspace:*",
