+++ apps/minifront/src/state/ibc-out.ts
@@ -17,7 +17,7 @@ import { toBaseUnit } from '@penumbra-zone/types/lo-hi';
 import { planBuildBroadcast } from './helpers';
 import { amountMoreThanBalance } from './send';
 import { getAssetId } from '@penumbra-zone/getters/metadata';
-import { assetPatterns } from '@penumbra-zone/constants/assets';
+import { assetPatterns } from '@penumbra-zone/types/assets';
 import { bech32, bech32m } from 'bech32';
 import { errorToast } from '@penumbra-zone/ui/lib/toast/presets';
 import { Chain } from '@penumbra-labs/registry';
