--- packages/constants/.eslintrc.cjs
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ['custom'],
-  parserOptions: {
-    project: true,
-    tsconfigRootDir: __dirname,
-  },
-};
