--- packages/constants/CHANGELOG.md
@@ -1,29 +1,0 @@
-# @penumbra-zone/constants
-
-## 4.0.0
-
-### Major Changes
-
-- 3148375: remove `/src/` path segment from exports
-
-## 3.0.0
-
-### Major Changes
-
-- 862283c: Using external registry for ibc chains
-
-## 2.0.0
-
-### Major Changes
-
-- 929d278: barrel imports to facilitate better tree shaking
-
-### Patch Changes
-
-- 8933117: Account for changes to core
-
-## 1.1.0
-
-### Minor Changes
-
-- Initial changest. Git tag v5.0.0 updates.
