--- packages/constants/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@penumbra-zone/constants",
-  "version": "4.0.0",
-  "license": "(MIT OR Apache-2.0)",
-  "type": "module",
-  "scripts": {
-    "build": "tsc && vite build",
-    "clean": "rm -rfv dist",
-    "lint": "eslint \"**/*.ts*\"",
-    "test": "vitest run"
-  },
-  "exports": {
-    "./assets": "./src/assets.ts"
-  },
-  "publishConfig": {
-    "files": [
-      "dist"
-    ],
-    "exports": {
-      "./assets": {
-        "import": "./dist/assets.js",
-        "types": "./dist/index.d.ts"
-      }
-    }
-  },
-  "devDependencies": {
-    "@buf/penumbra-zone_penumbra.bufbuild_es": "1.9.0-20240501193854-ea3cc743390d.1",
-    "@bufbuild/protobuf": "^1.9.0"
-  },
-  "peerDependencies": {
-    "@buf/penumbra-zone_penumbra.bufbuild_es": "1.9.0-20240501193854-ea3cc743390d.1",
-    "@bufbuild/protobuf": "^1.9.0"
-  }
-}
