--- packages/constants/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "tsconfig/base.json",
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "outDir": "dist",
-    "noEmit": true
-  }
-}
