+++ packages/getters/package.json
@@ -22,8 +22,7 @@
     }
   },
   "dependencies": {
-    "@penumbra-zone/bech32m": "workspace:*",
-    "@penumbra-zone/constants": "workspace:*"
+    "@penumbra-zone/bech32m": "workspace:*"
   },
   "devDependencies": {
     "@buf/penumbra-zone_penumbra.bufbuild_es": "1.9.0-20240501193854-ea3cc743390d.1",
