+++ packages/query/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@penumbra-zone/bech32m": "workspace:*",
-    "@penumbra-zone/constants": "workspace:*",
     "@penumbra-zone/crypto-web": "workspace:*",
     "@penumbra-zone/getters": "workspace:*",
     "@penumbra-zone/types": "workspace:*",
