+++ packages/services-context/package.json
@@ -15,7 +15,6 @@
   },
   "dependencies": {
     "@penumbra-labs/registry": "^7.0.0",
-    "@penumbra-zone/constants": "workspace:*",
     "@penumbra-zone/query": "workspace:*",
     "@penumbra-zone/storage": "workspace:*",
     "@penumbra-zone/types": "workspace:*",
