+++ packages/services/src/view-service/assets.ts
@@ -1,6 +1,6 @@
 import type { Impl } from '.';
 import { servicesCtx } from '../ctx/prax';
-import { assetPatterns, RegexMatcher } from '@penumbra-zone/constants/assets';
+import { assetPatterns, RegexMatcher } from '@penumbra-zone/types/assets';
 
 export const assets: Impl['assets'] = async function* (req, ctx) {
   const services = await ctx.values.get(servicesCtx)();
