+++ packages/types/package.json
@@ -24,7 +24,6 @@
   },
   "dependencies": {
     "@penumbra-zone/bech32m": "workspace:*",
-    "@penumbra-zone/constants": "workspace:*",
     "@penumbra-zone/getters": "workspace:*",
     "bignumber.js": "^9.1.2",
     "idb": "^8.0.0",
