+++ packages/ui/package.json
@@ -16,7 +16,6 @@
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.15.14",
     "@penumbra-zone/bech32m": "workspace:*",
-    "@penumbra-zone/constants": "workspace:*",
     "@penumbra-zone/getters": "workspace:*",
     "@penumbra-zone/perspective": "workspace:*",
     "@penumbra-zone/types": "workspace:*",
